add c14 to instant dapp package
