+++ packages/copilot/src/copilot/steps/topUp/TopUp.tsx
@@ -5,7 +5,7 @@
 import { useState } from "react";
 import { Intro } from "./Intro";
 import Onboard from "./Onboard";
-import C14 from "./C14";
+
 import LayerSwap from "./LayerSwap";
 import {
   CLICK_ON_DIFFERENT_ON_RAMP,
@@ -18,6 +18,7 @@ import { Squid } from "./Squid";
 import { raise } from "helpers";
 import Transak from "@evmosapps/instant-dapps/src/dapps/Transak";
 import { CypherD } from "@evmosapps/instant-dapps";
+import C14 from "@evmosapps/instant-dapps/src/dapps/C14";
 
 export const TopUp = () => {
   const [topUpType, setTopUpType] = useState("intro");
