+++ packages/instant-dapps/src/index.tsx
@@ -1,2 +1,3 @@
 export { CypherD } from "./dapps/CypherD";
 export * from "./dapps/Transak";
+export * from "./dapps/C14";
