+++ pages/dappstore/src/pages/dapp-explorer/partials/description-section.tsx
@@ -18,6 +18,7 @@ import { DescriptionItem } from "./description-item";
 import { DApp } from "../../../lib/fetch-explorer-data";
 // import { CypherD } from "@evmosapps/instant-dapps/src/dapps/CypherD";
 import Transak from "@evmosapps/instant-dapps/src/dapps/Transak";
+import C14 from "@evmosapps/instant-dapps/src/dapps/C14";
 export const DescriptiondApp = ({ dapp }: { dapp: DApp }) => {
   const imgId = dapp.cover?.id ?? dapp.icon?.id;
 
@@ -129,7 +130,7 @@ export const DescriptiondApp = ({ dapp }: { dapp: DApp }) => {
         <Frameline variant="secondary">
           <div className="flex items-center justify-center h-full">
             {/* TODO: add iframes */}
-            <Transak />
+            <C14 />
           </div>
         </Frameline>
       </div>
